backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
google-genai==0.2.2
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.6.1
